+++ src/test/java/org/ttzero/excel/entity/csv/StatementSheetTest.java
@@ -21,7 +21,6 @@ import org.ttzero.excel.Print;
 import org.ttzero.excel.entity.Column;
 import org.ttzero.excel.entity.ExcelWriteException;
 import org.ttzero.excel.entity.SQLWorkbookTest;
-import org.ttzero.excel.entity.Sheet;
 import org.ttzero.excel.entity.StatementSheet;
 import org.ttzero.excel.entity.Workbook;
 import org.ttzero.excel.entity.style.Fill;
@@ -83,7 +82,7 @@ public class StatementSheetTest extends SQLWorkbookTest {
                 .addSheet("select id, name, age from student"
                     , new Column("学号", int.class)
                     , new Column("姓名", String.class)
-                    , new Column("年龄", int.class, n -> n > 14 ? "高龄" : n)
+                    , new Column("年龄", int.class, n -> (int) n > 14 ? "高龄" : n)
                         .setStyleProcessor((o, style, sst) -> {
                             int n = (int) o;
                             if (n > 14) {
